refactor: 증정품 교환 로직을 별도의 객체로 분리
